requirements.txt
python-binance
tweepy
